usb: hub: Parallelize power-cycling of root-hub ports
Untill now we power-cycle (aka: disable power on a port
and re-enabling again) one port at a time.
Delay of 20ms for Port-power to change multiplies with
number of ports in this case.
So better we parallelize this process:
disable power on all ports, wait for port-power to stabilize
and then re-enable the power subsequently.

Backport XHCI specific code from top of tree U-Boot.
The commit is
http://git.denx.de/?p=u-boot.git;a=commit;h=0bf796f7ae22086f0504f3297e9fb4e96aa04161
    usb: hub: Parallelize power-cycling of root-hub ports

CRs-Fixed : 625723

Change-Id: If75f13d277f817f4e0ac79e85fa091504b5ce06d
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>
Signed-off-by: Sham Muthayyan <smuthayy@codeaurora.org>