ipq806x: Configure gsbi QUP to read from FIFO as long as clock is provided
While configuring gsbi QUP_MX_OUTPUT_COUNT
and QUP_MX_INPUT_COUNT for spi read a count of zero implies
read data as long as clock is supplied.
So configure it with zero and read as much data as we want.
This also helps us while reading data that is not aligned to
sector address and/or length.

CRs-Fixed: 808195

Change-Id: Ic78e0d39d260cae9ff51959a94a94d2c93af228e
Signed-off-by: Bharath Gandhi <bgandhi@codeaurora.org>