ipq806x: Fix MAC address fetch offset
The MAC adresses are written to the ART partition on a
zero-based index regardless of the number and/or index
of MACs applicable to the board. However, the driver
was picking them based on the MAC unit that was applicable
to the board. This resulted in the driver picking up
invalid MAC addresses.

For example on AP148, only GMAC 1 & 2 are applicable.
ART partition will have MAC address as follows

+-------+-------+-------+-------+
| MAC'1 | MAC'2 |  xxx  |  xxx  |
+-------+-------+-------+-------+
            ^       ^
            |       |
            |       +--- driver takes this for the second MAC
            |
            +--- driver takes this for the first MAC

Change the driver to pick the MAC as follows

+-------+-------+-------+-------+
| MAC'1 | MAC'2 |  xxx  |  xxx  |
+-------+-------+-------+-------+
    ^       ^
    |       |
    |       +--- driver takes this for the second MAC
    |
    +--- driver takes this for the first MAC

Change-Id: I49c177f2893ef4bfd3c9dfe68eb36b85a505777e
Signed-off-by: Varadarajan Narayanan <varada@codeaurora.org>