ipq806x: Added support for compressed dtb images
This commit lays a foundation for packing multiple dtb
images in a single itb image, by supporting compressed dtb
images. This is basically to support all boards across
different soc versions in a single itb image.
The plan is to compress the dtb images, so as to save flash space.

Change-Id: Ia8efee21501a0ff36c4da7ee6550bea81d36fd9c
Signed-off-by: Vasudevan Murugesan <vmuruges@codeaurora.org>