Support silent compile
U-Boot compilation generates too much output and it is
difficult to differentiate between the commands issued,
warnings/error messages etc. Introduce a silent compile
mode similar to Linux build. And, colorize the output
to make it easy on the eyes.

Set 'verbose' to a non-whitspace value to get the full
output.

Change-Id: I1ce39e6dde59473997cd1ea3844ec7261632b990
Signed-off-by: Varadarajan Narayanan <varada@codeaurora.org>