ipq806x: Added the flash crashdump data support
This patch adds the support for crashdump collection in
flash. A new command dumpipq_flash_data has been added for
this. It reads the TLV (Type Length Value) data passed
through kernel, extracts the required information and
writes to boot flash in the provided offset.

This change provides a reference to add any other
information needed to be saved.

Change-Id: I4d429359da6a29987b07da0f70b3f208eca0144d
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>