ipq806x: Enable cpu_type for macsec
cpu_type dts entry has been removed. Since we have the different
harware version images will be created from the openwrt. But the
cpu_type will be passed always as ipq8064. For macsec, they are
using the cpu_type as ipq8066. So we will have to enable the
cpu_type in u-boot and patched into dtb to access in the kernel.

Change-Id: Ic13b048fd5acdc0a65c27a02ae6c542a0477a5d5
Signed-off-by: Sham Muthayyan <smuthayy@codeaurora.org>