sysupgrade: Fix endianness problem when storing SW_ID to the disk
The problem was arised while calculating sw hash based on SW_ID from
the image. We are reading the SW_ID from the image and storing to disk
or file, This file is given as input to openssl command for generating
sw hash but due to endianness problem SW_ID bytes are reversed and hash
is getting calculated wrongly and causes sysupgrade failure.

In order to fix this, we will convert the SW_ID to big-endian and save
to the file so that it can help to calculate the proper sw hash.

Change-Id: If11aeb1f561f9785f3fe62c6a42c43a2d9215020
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>