Sysupgrade: Skip ubi section authentication in NOR/eMMC boot.
We have an image requirement which contains section names as hlos,
rootfs and ubi, which is used to perform upgrade in any boards,
sysupgrade fails in secure boot with below message in emmc boot.
"Error extracting /tmp/ubi.bin from ubi".
This error occurs when we try to authenticate ubi section in emmc boot.
So,this code-change helps to skip the authentication of ubi section and
proceed sysupgrade with remaining section as hlos and rootfs respectively.

Change-Id: Ib6817ad90630924cb2a7574cc79cc462665c159b
Signed-off-by: Avinash Pandey <avinasv@codeaurora.org>